@@ -1,129 +1,142 @@
 {
-   "name": "@stratio/egeo",
-   "version": "",
-   "description": "Egeo is an angular common components library",
-   "author": "Stratio",
-   "scripts": {
-      "ng": "ng",
-      "start": "gulp demo-app:copy:sourcecode && ng serve --project=egeo-demo",
-      "release": "gulp publish",
-      "clean": "gulp clean",
-      "build": "gulp prepare-build && npm run build:lib && npm run build:demo-app",
-      "build:web:aot": "gulp aot:build",
-      "publish": "gulp :publish:build-releases",
-      "build:lib": "ng build && npm run build:theme",
-      "build:theme": "gulp build",
-      "build:demo-app": "gulp demo-app:copy:sourcecode && ng build --project=egeo-demo --base-href=",
-      "test": "ng test --source-map=true --watch=false --code-coverage --progress=false --project=egeo",
-      "test:watch": "karma start test/karma.conf.js --auto-watch --no-single-run",
-      "lint": "ng lint",
-      "replace": "json -I -f package.json -e",
-      "commit": "git-cz",
-      "serve:coverage": "http-server target/html",
-      "serve:build": "npm run build:demo-app && http-server dist/demo-app"
-   },
-   "repository": {
-      "type": "git",
-      "url": "https://github.com/Stratio/egeo"
-   },
-   "license": "Apache-2.0",
-   "keywords": [
-      "angular",
-      "components",
-      "egeo",
-      "rollup",
-      "karma",
-      "gulp"
-   ],
-   "dependencies": {
-      "@angular/animations": "~7.2.15",
-      "@angular/common": "~7.2.15",
-      "@angular/compiler": "~7.2.15",
-      "@angular/core": "~7.2.15",
-      "@angular/forms": "~7.2.15",
-      "@angular/http": "~7.2.15",
-      "@angular/platform-browser": "~7.2.15",
-      "@angular/platform-browser-dynamic": "~7.2.15",
-      "@angular/router": "~7.2.15",
-      "core-js": "^2.5.4",
-      "lodash": "4.17.13",
-      "ngx-virtual-scroller": "3.0.2",
-      "reflect-metadata": "0.1.10",
-      "rxjs": "~6.3.3",
-      "zone.js": "~0.8.26",
-      "tslib": "1.7.1",
-      "prismjs": "1.16.0"
-   },
-   "devDependencies": {
-      "@angular-devkit/build-angular": "~0.10.0",
-      "@angular-devkit/build-ng-packagr": "~0.10.0",
-      "@angular/cli": "~7.3.9",
-      "@angular/compiler-cli": "~7.2.15",
-      "@angular/tsc-wrapped": "4.4.6",
-      "@types/chalk": "0.4.31",
-      "@types/fs-extra": "3.0.3",
-      "@types/glob": "5.0.30",
-      "@types/gulp": "3.8.32",
-      "@types/jasmine": "2.8.6",
-      "@types/jasminewd2": "~2.0.3",
-      "@types/js-beautify": "0.0.30",
-      "@types/lodash": "4.14.136",
-      "@types/minimist": "1.2.0",
-      "@types/node": "~8.9.4",
-      "@types/prismjs": "1.16.0",
-      "@types/run-sequence": "0.0.29",
-      "@types/showdown": "1.4.32",
-      "autoprefixer": "7.1.1",
-      "chalk": "2.1.0",
-      "commitizen": "2.9.6",
-      "cssnano": "^4.1.7",
-      "fs-extra": "3.0.1",
-      "gulp": "3.9.1",
-      "gulp-clean": "0.3.2",
-      "gulp-clean-css": "^4.0.0",
-      "gulp-connect": "5.0.0",
-      "gulp-htmlmin": "3.0.0",
-      "gulp-if": "2.0.2",
-      "gulp-postcss": "^8.0.0",
-      "gulp-remove-html-comments": "1.0.1",
-      "gulp-sass": "^4.0.2",
-      "gulp-sourcemaps": "2.6.0",
-      "gulp-strip-comments": "^2.5.2",
-      "http-server": "^0.11.1",
-      "intl": "1.2.5",
-      "jasmine-core": "~2.99.1",
-      "jasmine-spec-reporter": "~4.2.1",
-      "js-beautify": "1.6.14",
-      "json": "9.0.6",
-      "karma": "~3.0.0",
-      "karma-coverage-istanbul-reporter": "2.0.4",
-      "karma-jasmine": "1.1.1",
-      "karma-junit-reporter": "1.2.0",
-      "karma-mocha-reporter": "2.2.5",
-      "karma-phantomjs-launcher": "1.0.4",
-      "karma-typescript": "3.0.12",
-      "karma-typescript-angular2-transform": "1.0.2",
-      "madge": "1.6.0",
-      "markdown-to-html-pipe": "1.2.4",
-      "minimist": "1.2.0",
-      "ng-packagr": "^4.4.0",
-      "resolve-bin": "0.4.0",
-      "run-sequence": "^2.2.1",
-      "scss-bundle": "2.4.0",
-      "showdown": "1.8.2",
-      "sorcery": "0.10.0",
-      "stylelint": "8.3.0",
-      "ts-node": "~7.0.0",
-      "tsconfig-paths": "2.2.0",
-      "tslint": "~5.11.0",
-      "tsickle": "0.29.0",
-      "typescript": "~3.1.1",
-      "uglify-js": "2.8.22",
-      "word-wrap": "1.2.3"
-   },
-   "config": {
-      "commitizen": {
-         "path": "./tools/commitizen-format/st-commitizen.js"
-      }
-   }
+  "name": "@stratio/egeo",
+  "version": "",
+  "description": "Egeo is an angular common components library",
+  "author": "Stratio",
+  "scripts": {
+    "ng": "ng",
+    "start": "gulp demo-app:copy:sourcecode && ng serve --project=egeo-demo",
+    "release": "gulp publish",
+    "clean": "gulp clean",
+    "build": "gulp prepare-build && npm run build:lib && npm run build:demo-app",
+    "build:elements": "cd projects/egeo-elements && sh buildElements.sh",
+    "build:web:aot": "gulp aot:build",
+    "publish": "gulp :publish:build-releases",
+    "build:lib": "ng build && npm run build:theme",
+    "build:theme": "gulp build",
+    "build:demo-app": "gulp demo-app:copy:sourcecode && ng build --project=egeo-demo --base-href=",
+    "test": "ng test --source-map=true --watch=false --code-coverage --progress=false --project=egeo",
+    "test:watch": "karma start test/karma.conf.js --auto-watch --no-single-run",
+    "lint": "ng lint",
+    "replace": "json -I -f package.json -e",
+    "commit": "git-cz",
+    "serve:coverage": "http-server target/html",
+    "serve:build": "npm run build:demo-app && http-server dist/demo-app"
+  },
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/Stratio/egeo"
+  },
+  "license": "Apache-2.0",
+  "keywords": [
+    "angular",
+    "components",
+    "egeo",
+    "rollup",
+    "karma",
+    "gulp"
+  ],
+  "dependencies": {
+    "@angular-builders/custom-webpack": "~8.0.2",
+    "@angular/animations": "~8.2.14",
+    "@angular/common": "~8.2.14",
+    "@angular/compiler": "~8.2.14",
+    "@angular/core": "~8.2.14",
+    "@angular/elements": "~8.2.14",
+    "@angular/forms": "~8.2.14",
+    "@angular/platform-browser": "~8.2.14",
+    "@angular/platform-browser-dynamic": "~8.2.14",
+    "@angular/router": "~8.2.14",
+    "@webcomponents/webcomponentsjs": "~2.4.0",
+    "classlist.js": "^1.1.20150312",
+    "core-js": "^2.5.4",
+    "document-register-element": "~1.8.1",
+    "lodash": "4.17.13",
+    "ngx-build-plus": "~9.0.2",
+    "ngx-virtual-scroller": "3.0.2",
+    "reflect-metadata": "0.1.10",
+    "rxjs": "~6.5.3",
+    "zone.js": "~0.9.1",
+    "tslib": "1.10.0",
+    "prismjs": "1.16.0"
+  },
+  "devDependencies": {
+    "@angular-devkit/build-angular": "~0.803.20",
+    "@angular-devkit/build-ng-packagr": "~0.803.20",
+    "@angular/cli": "~8.3.20",
+    "@angular/compiler-cli": "~8.2.14",
+    "@angular/tsc-wrapped": "4.4.6",
+    "@types/chalk": "0.4.31",
+    "@types/fs-extra": "3.0.3",
+    "@types/glob": "5.0.30",
+    "@types/gulp": "3.8.32",
+    "@types/jasmine": "2.8.6",
+    "@types/jasminewd2": "~2.0.3",
+    "@types/js-beautify": "0.0.30",
+    "@types/lodash": "4.14.136",
+    "@types/minimist": "1.2.0",
+    "@types/node": "~8.9.4",
+    "@types/prismjs": "1.16.0",
+    "@types/run-sequence": "0.0.29",
+    "@types/showdown": "1.4.32",
+    "autoprefixer": "7.1.1",
+    "caniuse-lite": "1.0.30000997",
+    "chalk": "2.1.0",
+    "codelyzer": "~5.0.1",
+    "commitizen": "2.9.6",
+    "compression-webpack-plugin": "3.0.1",
+    "cssnano": "^4.1.7",
+    "fs-extra": "3.0.1",
+    "gulp": "3.9.1",
+    "gulp-clean": "0.3.2",
+    "gulp-clean-css": "^4.0.0",
+    "gulp-connect": "5.0.0",
+    "gulp-htmlmin": "3.0.0",
+    "gulp-if": "2.0.2",
+    "gulp-postcss": "^8.0.0",
+    "gulp-remove-html-comments": "1.0.1",
+    "gulp-sass": "^4.0.2",
+    "gulp-sourcemaps": "2.6.0",
+    "gulp-strip-comments": "^2.5.2",
+    "http-server": "^0.11.1",
+    "intl": "1.2.5",
+    "jasmine-core": "~2.99.1",
+    "jasmine-spec-reporter": "~4.2.1",
+    "js-beautify": "1.6.14",
+    "json": "9.0.6",
+    "karma": "~3.0.0",
+    "karma-coverage-istanbul-reporter": "2.0.4",
+    "karma-jasmine": "1.1.1",
+    "karma-junit-reporter": "1.2.0",
+    "karma-mocha-reporter": "2.2.5",
+    "karma-phantomjs-launcher": "1.0.4",
+    "karma-typescript": "3.0.12",
+    "karma-typescript-angular2-transform": "1.0.2",
+    "madge": "1.6.0",
+    "markdown-to-html-pipe": "1.2.4",
+    "minimist": "1.2.0",
+    "ng-packagr": "^5.1.0",
+    "node-sass": "~4.13.0",
+    "resolve-bin": "0.4.0",
+    "run-sequence": "^2.2.1",
+    "scss-bundle": "2.4.0",
+    "showdown": "1.8.2",
+    "sorcery": "0.10.0",
+    "stylelint": "8.3.0",
+    "terser-webpack-plugin": "2.2.3",
+    "ts-node": "~7.0.0",
+    "tsconfig-paths": "2.2.0",
+    "tsickle": "~0.37.0",
+    "tslint": "~5.11.0",
+    "typescript": "~3.5.3",
+    "uglify-js": "2.8.22",
+    "uglifyjs-webpack-plugin": "1.1.0",
+    "uuid": "~3.3.2",
+    "word-wrap": "1.2.3"
+  },
+  "config": {
+    "commitizen": {
+      "path": "./tools/commitizen-format/st-commitizen.js"
+    }
+  }
 }
