@@ -68,7 +68,7 @@ export class StPageTitleComponent implements ControlValueAccessor, OnInit {
    @Output() clickButton: EventEmitter<any> = new EventEmitter();
    @Output() edit: EventEmitter<string> = new EventEmitter();
 
-   @ViewChild('input') input: any;
+   @ViewChild('input', {static: false}) input: any;
 
    public focus: boolean = false;
    public hasPreTitleValue: boolean = false;
