@@ -29,7 +29,7 @@ import { StDemoLoggerService } from '../shared/st-demo-logger/st-demo-logger.ser
 })
 
 export class StModalDemoComponent implements AfterViewInit {
-   @ViewChild('loadModal', { read: ViewContainerRef }) target: ViewContainerRef;
+   @ViewChild('loadModal', { read: ViewContainerRef, static: false }) target: ViewContainerRef;
    public configDoc: any = {
       html: 'demo/st-modal-demo/st-modal-demo.component.html',
       ts: 'demo/st-modal-demo/st-modal-demo.component.ts',
