@@ -45,6 +45,7 @@ export { StClickOutside } from './directives/st-click-outside/st-click-outside.d
 
 // Alerts
 export { StAlertsModule } from './st-alerts/st-alerts.module';
+export { StAlertsComponent } from './st-alerts/st-alerts.component';
 export { StAlertLink, STALERT_SEVERITY } from './st-alerts/st-alerts.model';
 export { StAlertsService } from './st-alerts/st-alerts.service';
 
@@ -54,25 +55,27 @@ export {
    StBreadCrumbMode
 } from './st-breadcrumbs/st-breadcrumbs.interface';
 export { StBreadcrumbsModule } from './st-breadcrumbs/st-breadcrumbs.module';
+export { StBreadCrumbsComponent } from './st-breadcrumbs/st-breadcrumbs.component';
 
 // Bubble
 export { StBubbleModule } from './st-bubble/st-bubble.module';
+export { StBubbleComponent } from './st-bubble/st-bubble.component';
 
 // Color picker
 export { StColorPickerModule } from './st-color-picker/st-color-picker.module';
+export { StColorPickerComponent } from './st-color-picker/st-color-picker.component';
 
 // Checkbox
 export { StCheckboxModule } from './st-checkbox/st-checkbox.module';
+export { StCheckboxComponent } from './st-checkbox/st-checkbox.component';
 
 // Docs
-export {
-   StDocsModule
-} from './st-docs/st-docs.module';
+export { StDocsModule } from './st-docs/st-docs.module';
+export { StDocsComponent } from './st-docs/st-docs.component';
 
 // Draggable car
-export {
-   StDraggableCardModule
-} from './st-draggable-card/st-draggable-card.module';
+export { StDraggableCardModule } from './st-draggable-card/st-draggable-card.module';
+export { StDraggableCardComponent } from './st-draggable-card/st-draggable-card.component';
 
 // Dropdown menu
 export {
@@ -83,20 +86,25 @@ export {
    StDropDownMenuItemSchema
 } from './st-dropdown-menu/st-dropdown-menu.interface';
 export { StDropdownMenuModule } from './st-dropdown-menu/st-dropdown-menu.module';
+export { StDropdownMenuComponent } from './st-dropdown-menu/st-dropdown-menu.component';
 
 // Filter selector
 export { StFilterSelectorModule } from './st-filter-selector/st-filter-selector.module';
+export { StFilterSelectorComponent } from './st-filter-selector/st-filter-selector.component';
 
 // File button
 export { StFileButtonModule } from './st-file-button/st-file-button.module';
+export { StFileButtonComponent } from './st-file-button/st-file-button.component';
 
 // Footers
 export { StFooterModule } from './st-footer/st-footer.module';
+export { StFooterComponent } from './st-footer/st-footer.component';
 export { StFooterLink } from './st-footer/st-footer.model';
 
 // Foreground notifications
 
 export { StForegroundNotificationsModule } from './st-foreground-notifications/st-foreground-notifications.module';
+export { StForegroundNotificationsComponent } from './st-foreground-notifications/st-foreground-notifications';
 
 // Header
 export {
@@ -106,6 +114,7 @@ export {
    StHeaderSubMenuOptionSchema
 } from './st-header/st-header.model';
 export { StHeaderModule } from './st-header/st-header.module';
+export { StHeaderComponent } from './st-header/st-header.component';
 
 // Dynamic form
 export { StFormModule } from './st-form/st-form.module';
@@ -118,11 +127,11 @@ export {StNotificationElement} from './st-foreground-notifications/st-foreground
 
 // Fullscreen layout
 export { StFullscreenLayoutModule } from './st-fullscreen-layout/st-fullscreen-layout.module';
+export { StFullscreenLayoutComponent } from './st-fullscreen-layout/st-fullscreen-layout';
 
 // Horizontal Tabs
-export {
-   StHorizontalTabsModule
-} from './st-horizontal-tabs/st-horizontal-tabs.module';
+export { StHorizontalTabsModule } from './st-horizontal-tabs/st-horizontal-tabs.module';
+export { StHorizontalTabsComponent } from './st-horizontal-tabs/st-horizontal-tabs.component';
 export {
    StHorizontalTabStatus,
    StHorizontalTab,
@@ -131,9 +140,11 @@ export {
 
 // Info box
 export { StInfoBoxModule } from './st-info-box/st-info-box.module';
+export { StInfoBoxComponent } from './st-info-box/st-info-box.component';
 
 // Input
 export { StInputModule } from './st-input/st-input.module';
+export { StInputComponent } from './st-input/st-input.component';
 export {
    StInputError,
    StInputErrorSchema
@@ -144,6 +155,9 @@ export {
    StItemListModule
 } from './st-item-list/st-item-list.module';
 export {
+   StItemListComponent
+} from './st-item-list/st-item-list.component';
+export {
    StItemListElement,
    StItemListConfig,
    StItemListConfigSchema
@@ -151,6 +165,7 @@ export {
 
 // Launcher
 export { StLauncherModule } from './st-launcher/st-launcher.module';
+export { StLauncherComponent } from './st-launcher/st-launcher.component';
 export {
    StLauncherItem,
    StLauncherItemSchema,
@@ -160,6 +175,7 @@ export {
 
 // Menu
 export { StMenuModule } from './st-menu/st-menu.module';
+export { StMenuComponent } from './st-menu/st-menu.component';
 export {
    StMenuModel,
    StMenuStatus
@@ -174,6 +190,7 @@ export {
 } from './st-modal/st-modal.model';
 export { StModalService } from './st-modal/st-modal.service';
 export { StModalModule } from './st-modal/st-modal.module';
+export { StModalComponent } from './st-modal/st-modal.component';
 
 // Modal2
 export { StModal2Component } from './st-modal2/st-modal2.component';
@@ -181,6 +198,7 @@ export { StModal2Module } from './st-modal2/st-modal2.module';
 
 // Page Title
 export { StPageTitleModule } from './st-page-title/st-page-title.module';
+export { StPageTitleComponent } from './st-page-title/st-page-title.component';
 
 // Pagination
 export {
@@ -192,9 +210,11 @@ export {
 export { StPaginationPipe } from './st-pagination/st-pagination.pipe';
 export { StPaginationService } from './st-pagination/st-pagination.service';
 export { StPaginationModule } from './st-pagination/st-pagination.module';
+export { StPaginationComponent } from './st-pagination/st-pagination.component';
 
 // Pop Over
 export { StPopOverModule } from './st-pop-over/st-pop-over.module';
+export { StPopOverComponent } from './st-pop-over/st-pop-over.component';
 
 // Prism
 export {
@@ -204,13 +224,17 @@ export {
 // Progress bar
 
 export { StProgressBarModule } from './st-progress-bar/st-progress-bar.module';
+export { StProgressBarComponent } from './st-progress-bar/st-progress-bar';
+
 
 // Radio
 export { StRadioModule } from './st-radio/st-radio.module';
+export { StRadioComponent } from './st-radio/st-radio.component';
 export { RadioChange } from './st-radio/st-radio.change';
 
 // Radio menu
 export { StRadioMenuModule } from './st-radio-menu/st-radio-menu.module';
+export { StRadioMenuComponent } from './st-radio-menu/st-radio-menu.component';
 export {
    StRadioMenuOption,
    StRadioMenuOptionSchema
@@ -218,6 +242,7 @@ export {
 
 // Search
 export { StSearchModule } from './st-search/st-search.module';
+export { StSearchComponent } from './st-search/st-search.component';
 export {
    StSearchEvent,
    StSearchEventOrigin
@@ -225,25 +250,31 @@ export {
 
 // Select
 export { StSelectModule } from './st-select/st-select.module';
+export { StSelectComponent } from './st-select/st-select';
 
 // Sidebar
 export { StSidebarModule } from './st-sidebar/st-sidebar.module';
+export { StSidebarComponent } from './st-sidebar/st-sidebar.component';
 export { StSidebarItem } from './st-sidebar/st-sidebar-item.interface';
 export { StSidebarVisualMode } from './st-sidebar/st-sidebar-visual-mode';
 
 // Spinner
 export { StSpinnerModule } from './st-spinner/st-spinner.module';
+export { StSpinnerComponent } from './st-spinner/st-spinner.component';
 
 // Switch
 export { StSwitchModule } from './st-switch/st-switch.module';
+export { StSwitchComponent } from './st-switch/st-switch.component';
 
 // Table
 export { StTableModule } from './st-table/st-table.module';
+export { StTableComponent } from './st-table/st-table.component';
 export { StTableHeader } from './st-table/shared/table-header.interface';
 export { Order, ORDER_TYPE } from './st-table/shared/order';
 
 // Tag Input
 export { StTagInputModule } from './st-tag-input/st-tag-input.module';
+export { StTagInputComponent } from './st-tag-input/st-tag-input.component';
 
 // Textarea
 export {
@@ -251,6 +282,7 @@ export {
    StTextareaErrorSchema
 } from './st-textarea/st-textarea.error.model';
 export { StTextareaModule } from './st-textarea/st-textarea.module';
+export { StTextareaComponent } from './st-textarea/st-textarea.component';
 
 // Toogle buttons
 export {
@@ -263,11 +295,13 @@ export {
 
 // Tooltip
 export { StTooltipModule } from './st-tooltip/st-tooltip.module';
+export { StTooltipComponent } from './st-tooltip/st-tooltip.component';
 
 // Two list selection
 export {
    StTwoListSelectionModule
 } from './st-two-list-selection/st-two-list-selection.module';
+export { StTwoListSelectionComponent} from './st-two-list-selection/st-two-list-selection.component';
 export {
    StTwoListSelectionConfig,
    StTwoListSelectionConfigSchema,
@@ -281,15 +315,19 @@ export {
 
 // Vertical tabs
 export { StVerticalTabsModule } from './st-vertical-tabs/st-vertical-tabs.module';
+export { StVerticalTabsComponent  } from './st-vertical-tabs/st-vertical-tabs.component';
 
 // Widget
 export { StWidgetModule } from './st-widget/st-widget.module';
+export { StWidgetComponent } from './st-widget/st-widget.component';
 
 
 // Zero page
 export { StZeroPageModule } from './st-zero-page/st-zero-page.module';
+export { StZeroPageComponent } from './st-zero-page/st-zero-page.component';
 
 export { StPopModule } from './st-pop/st-pop.module';
+export { StPopComponent } from './st-pop/st-pop.component';
 export { StPopOffset, StPopPlacement } from './st-pop/st-pop.model';
 
 // Utils
