@@ -10,7 +10,7 @@
  */
 import { DebugElement, SimpleChange, SimpleChanges, NO_ERRORS_SCHEMA } from '@angular/core';
 import { async, ComponentFixture, fakeAsync, TestBed, tick } from '@angular/core/testing';
-import { Http } from '@angular/http';
+
 import { By } from '@angular/platform-browser';
 import { VirtualScrollerModule } from 'ngx-virtual-scroller';
 import {
